Time: 1 ms (98.23%), Space: 43.5 MB (36.89%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.java
@@ -2,7 +2,7 @@ class Solution {
     public int scoreOfString(String s) {
         int result =0;
         for(int i=0;i<s.length()-1;i++){
-            result += Math.abs((int)s.charAt(i) - (int)s.charAt(i+1));
+            result += Math.abs(s.charAt(i) - s.charAt(i+1));
         }
         return result;
     }
